Separate ionospheric screen functions
