+++ tests/processing_components/test_ionospheric_screen.py
@@ -9,23 +9,28 @@ import astropy.units as u
 import numpy
 from astropy.coordinates import SkyCoord
 
-from data_models.memory_data_models import SkyModel, Image
+from data_models.parameters import arl_path
 from data_models.polarisation import PolarisationFrame
+from data_models.memory_data_models import SkyModel
+from processing_components.calibration.operations import qa_gaintable
+from processing_components.image.operations import import_image_from_fits
+from processing_components.imaging.base import create_image_from_visibility
 from processing_components.imaging.primary_beams import create_low_test_beam
 from processing_components.simulation.mpc import expand_skymodel_by_skycomponents
+from processing_components.simulation.ionospheric_screen import create_gaintable_from_screen
 from processing_components.simulation.testing_support import create_named_configuration, \
     create_low_test_skycomponents_from_gleam
 from processing_components.simulation.testing_support import create_test_image
 from processing_components.skycomponent.operations import apply_beam_to_skycomponent, remove_neighbouring_components
 from processing_components.skycomponent.operations import filter_skycomponents_by_flux
-from processing_components.skymodel.operations import image_voronoi_iter
 from processing_components.visibility.base import create_blockvisibility
+from processing_components.skymodel.operations import image_voronoi_iter
 from processing_library.image.operations import create_image
 
 log = logging.getLogger(__name__)
 
 
-class TestMPC(unittest.TestCase):
+class TestIonosphericScreen(unittest.TestCase):
     def setUp(self):
         from data_models.parameters import arl_path
         dec = -40.0 * u.deg
@@ -33,7 +38,7 @@ class TestMPC(unittest.TestCase):
         self.lowcore = create_named_configuration('LOWBD2', rmax=300.0)
         self.dir = arl_path('test_results')
         self.times = numpy.linspace(-10.0, 10.0, 3) * numpy.pi / (3600.0 * 12.0)
-        
+
         self.frequency = numpy.array([1e8])
         self.channel_bandwidth = numpy.array([1e6])
         self.phasecentre = SkyCoord(ra=+0.0 * u.deg, dec=dec, frame='icrs', equinox='J2000')
@@ -47,8 +52,14 @@ class TestMPC(unittest.TestCase):
         self.model = create_image(npixel=512, cellsize=0.000015, polarisation_frame=PolarisationFrame("stokesI"),
                                   frequency=self.frequency, channel_bandwidth=self.channel_bandwidth,
                                   phasecentre=self.phasecentre)
+
+    def test_read_screen(self):
+        screen = import_image_from_fits(arl_path('data/models/test_mpc_screen.fits'))
+        assert screen.data.shape == (1, 3, 2000, 2000), screen.data.shape
     
-    def test_expand_skymodel_by_skycomponents(self):
+    def test_create_gaintable_from_screen(self):
+
+        screen = import_image_from_fits(arl_path('data/models/test_mpc_screen.fits'))
         beam = create_test_image(cellsize=0.0015, phasecentre=self.vis.phasecentre,
                                  frequency=self.frequency)
         
@@ -65,46 +76,6 @@ class TestMPC(unittest.TestCase):
         
         actual_components = filter_skycomponents_by_flux(pb_gleam_components, flux_min=1.0)
         
-        assert len(actual_components) == 15, len(actual_components)
-        sm = SkyModel(image=self.model, components=actual_components)
-        assert len(sm.components) == len(actual_components)
-        
-        scatter_sm = expand_skymodel_by_skycomponents(sm)
-        assert len(scatter_sm) == len(actual_components)
-        assert len(scatter_sm[0].components) == 1
-    
-    def test_expand_skymodel_voronoi(self):
-        self.model = create_image(npixel=256, cellsize=0.001, polarisation_frame=PolarisationFrame("stokesI"),
-                                  frequency=self.frequency, channel_bandwidth=self.channel_bandwidth,
-                                  phasecentre=self.phasecentre)
-        beam = create_low_test_beam(self.model)
-        
-        gleam_components = create_low_test_skycomponents_from_gleam(flux_limit=1.0,
-                                                                    phasecentre=self.phasecentre,
-                                                                    frequency=self.frequency,
-                                                                    polarisation_frame=PolarisationFrame(
-                                                                        'stokesI'),
-                                                                    radius=0.1)
-        
-        pb_gleam_components = apply_beam_to_skycomponent(gleam_components, beam)
-        
-        actual_components = filter_skycomponents_by_flux(pb_gleam_components, flux_min=1.0)
-        _, actual_components = remove_neighbouring_components(actual_components, 0.05)
-        
-        for imask, mask in enumerate(image_voronoi_iter(self.model, actual_components)):
-            mask.data *= beam.data
-            assert isinstance(mask, Image)
-            assert mask.data.dtype == "float"
-            assert numpy.sum(mask.data) > 1
-            # import matplotlib.pyplot as plt
-            # from processing_components.image.operations import show_image
-            # show_image(mask)
-            # plt.show()
-
-        assert len(actual_components) == 6, len(actual_components)
-        sm = SkyModel(image=self.model, components=actual_components)
-        assert len(sm.components) == len(actual_components)
-        
-        scatter_sm = expand_skymodel_by_skycomponents(sm)
-        assert len(scatter_sm) == len(actual_components)
-        assert len(scatter_sm[0].components) == 1
+        gaintables = create_gaintable_from_screen(self.vis, actual_components, screen)
+        assert len(gaintables) == len(actual_components), len(gaintables)
+        assert gaintables[0].gain.shape == (3, 94, 1, 1, 1), gaintables[0].gain.shape
