+++ wrappers/serial/simulation/mpc.py
@@ -2,5 +2,6 @@
 Functions that aid testing in various ways. 
 """
 
-from processing_components.simulation.mpc import find_pierce_points, create_gaintable_from_screen, \
-    expand_skymodel_by_skycomponents, sum_visibility_over_partitions, calculate_sf_from_screen
+from processing_components.simulation.mpc import expand_skymodel_by_skycomponents, sum_visibility_over_partitions
+from processing_components.simulation.ionospheric_screen import find_pierce_points, create_gaintable_from_screen, \
+    calculate_sf_from_screen
